Cleaning after testsCI test error fix - file split
